Use float64 for RUv2 metrics calculations
Convert RU v2 related values from int64 to float64 across the codebase to preserve fractional RU values and avoid truncation. Key changes: RUV2 metrics CalculateRUValues/TotalRU now return float64; FormatRUV2Metrics prints floats with two decimals; runtime stats, slow log, resultset tracker, and statement summary fields updated to float64; info schema columns for RequestUnitV2 changed to DOUBLE; tests and summary reader/column factories adjusted accordingly. This ensures consistent handling and formatting of RUv2 totals (TiDB/TiKV/TiFlash) throughout the system.

+++ pkg/executor/adapter.go
@@ -1733,11 +1733,11 @@ func (a *ExecStmt) finalizeStatementRUV2Metrics() {
 	}
 }
 
-func calculateStatementTotalRUV2(metrics *execdetails.RUV2Metrics, weights execdetails.RUV2Weights, ruDetail *util.RUDetails) int64 {
-	var tiKVRU, tiFlashRU int64
+func calculateStatementTotalRUV2(metrics *execdetails.RUV2Metrics, weights execdetails.RUV2Weights, ruDetail *util.RUDetails) float64 {
+	var tiKVRU, tiFlashRU float64
 	if ruDetail != nil {
-		tiKVRU = int64(ruDetail.TiKVRUV2())
-		tiFlashRU = int64(ruDetail.TiflashRU())
+		tiKVRU = ruDetail.TiKVRUV2()
+		tiFlashRU = ruDetail.TiflashRU()
 	}
 	if metrics == nil {
 		return tiKVRU + tiFlashRU
